chore: Small correction for time to be an integer conversion and not a float

+++ tests/orm/io/exporters/test_section_reliability_exporter.py
@@ -94,6 +94,6 @@ class TestSectionReliabilityExporter:
                     col_idx
                 ]
             )
-            assert _orm_assessment.time == float(
+            assert _orm_assessment.time == int(
                 section_reliability_with_values.SectionReliability.columns[col_idx]
             )
